+++ apps/log-server/src/logs/logs.controller.ts
@@ -24,14 +24,19 @@ export class LogsController {
     @EnvId() envId: string,
     @Body() body: CreateLogInput,
   ): Promise<{}> {
-    this.logger.log(`Received log for ${envId}`);
-    const queueInput: QLogsInput = {
-      envId,
-      trace: body.trace,
-      log: body.log,
-    };
-    await this.logsQueue.add(queueInput);
-    this.logger.log(`Added ${envId} to queue`);
-    return res.status(200).json({});
+    try {
+      this.logger.log(`Received log for ${envId}`);
+      const queueInput: QLogsInput = {
+        envId,
+        trace: body.trace,
+        log: body.log,
+      };
+      await this.logsQueue.add(queueInput);
+      this.logger.log(`Added ${envId} to queue`);
+      return res.status(200).json({});
+    } catch(e) {
+      this.logger.error(e);
+      throw e;
+    }
   }
 }