+++ apps/montelo/rollup.config.cjs
@@ -1,6 +1,8 @@
 const resolve = require("@rollup/plugin-node-resolve").nodeResolve;
 const commonjs = require("@rollup/plugin-commonjs");
 const typescript = require("rollup-plugin-typescript2");
+const json = require("@rollup/plugin-json");
+const wasm = require("@rollup/plugin-wasm");
 
 /**
  * @type {import("rollup").RollupOptions[]}
@@ -10,17 +12,17 @@ module.exports = [
     input: "src/index.ts",
     output: {
       dir: "dist/bundle.js",
-      format: "commonjs"
+      format: "commonjs",
     },
-    plugins: [typescript(), resolve(), commonjs()]
+    plugins: [typescript(), resolve(), commonjs(), json(), wasm()],
   },
   // Separate configuration for the CLI script
   {
     input: "../../packages/cli/src/cli.ts",
     output: {
       file: "dist/cli.js",
-      format: "commonjs"
+      format: "commonjs",
     },
-    plugins: [typescript()]
-  }
+    plugins: [typescript()],
+  },
 ];