+++ packages/browser-client/src/models/LogDto.ts
@@ -54,7 +54,7 @@ export interface LogDto {
      * @type {string}
      * @memberof LogDto
      */
-    type: LogDtoTypeEnum;
+    source: LogDtoSourceEnum;
     /**
      * 
      * @type {string}
@@ -145,11 +145,14 @@ export interface LogDto {
 /**
  * @export
  */
-export const LogDtoTypeEnum = {
+export const LogDtoSourceEnum = {
     Manual: 'MANUAL',
-    Openai: 'OPENAI'
+    Openai: 'OPENAI',
+    Anthropic: 'ANTHROPIC',
+    Mistral: 'MISTRAL',
+    Cohere: 'COHERE'
 } as const;
-export type LogDtoTypeEnum = typeof LogDtoTypeEnum[keyof typeof LogDtoTypeEnum];
+export type LogDtoSourceEnum = typeof LogDtoSourceEnum[keyof typeof LogDtoSourceEnum];
 
 
 /**
@@ -162,7 +165,7 @@ export function instanceOfLogDto(value: object): boolean {
     isInstance = isInstance && "envId" in value;
     isInstance = isInstance && "parentLogId" in value;
     isInstance = isInstance && "name" in value;
-    isInstance = isInstance && "type" in value;
+    isInstance = isInstance && "source" in value;
     isInstance = isInstance && "model" in value;
     isInstance = isInstance && "input" in value;
     isInstance = isInstance && "output" in value;
@@ -196,7 +199,7 @@ export function LogDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): Lo
         'envId': json['envId'],
         'parentLogId': json['parentLogId'],
         'name': json['name'],
-        'type': json['type'],
+        'source': json['source'],
         'model': json['model'],
         'input': json['input'],
         'output': json['output'],
@@ -228,7 +231,7 @@ export function LogDtoToJSON(value?: LogDto | null): any {
         'envId': value.envId,
         'parentLogId': value.parentLogId,
         'name': value.name,
-        'type': value.type,
+        'source': value.source,
         'model': value.model,
         'input': value.input,
         'output': value.output,