+++ packages/promptfile/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @montelo/promptfile
 
+## 1.1.48
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/common@1.1.48
+
+## 1.1.47
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/common@1.1.47
+
+## 1.1.46
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/common@1.1.46
+
+## 1.1.45
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/common@1.1.45
+
+## 1.1.44
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/common@1.1.44
+
+## 1.1.43
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/common@1.1.43
+
+## 1.1.42
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/common@1.1.42
+
 ## 1.1.41
 
 ### Patch Changes